#16 Hiding jersey errors for specific tests when running maven build.
I've added an extension that can be used to disable Jersey's Errors for the tests that cause internal errors to happen.

Errors are only disabled in the main build which means if you are doing active development you will still see them (in case they are needed) and will need to deliberately disable any new exceptions.